refactor: Timezone validation and proper validation for chat endpoint
