+++ packages/@n8n/api-types/src/chat-hub.ts
@@ -315,26 +315,8 @@ export const chatAttachmentSchema = z.object({
 	fileName: z.string(),
 });
 
-export const isValidTimeZone = (tz: string): boolean => {
-	try {
-		// Throws if invalid timezone
-		new Intl.DateTimeFormat('en-US', { timeZone: tz });
-		return true;
-	} catch {
-		return false;
-	}
-};
-
-export const StrictTimeZoneSchema = z
-	.string()
-	.min(1)
-	.max(50)
-	.regex(/^[A-Za-z0-9_/+-]+$/)
-	.refine(isValidTimeZone, {
-		message: 'Unknown or invalid time zone',
-	});
-
-export const TimeZoneSchema = StrictTimeZoneSchema.optional().catch(undefined);
+import { TimeZoneSchema } from './schemas/timezone.schema';
+export { isValidTimeZone, StrictTimeZoneSchema, TimeZoneSchema } from './schemas/timezone.schema';
 
 export type ChatAttachment = z.infer<typeof chatAttachmentSchema>;
